file renaming to all camelCase, reorganizing models for redux saga
