--- server/models/UserResource.ts
@@ -1,14 +1,0 @@
-export interface PacamaraUser {
-  id: number;
-  username: string;
-  profile_pic?: string;
-  methods_default_id?: number;
-  kettle?: string;
-  grinder?: string;
-  tds_min?: number;
-  tds_max?: number;
-  ext_min?: number;
-  ext_max?: number;
-  name?: string;
-  methods_default_lrr?: number;
-}
