+++ server/routes/brewsRouter.ts
@@ -1,6 +1,6 @@
 import express, { Request, Response } from 'express';
 import pool from '../modules/pool';
-import { rejectUnauthenticated } from '../modules/authentication.middleware';
+import { rejectUnauthenticated } from '../modules/authenticationMiddleware';
 const router: express.Router = express.Router();
 
 // GET route to grab all brew instances for a coffee
