+++ server/routes/flavorsRouter.ts
