+++ server/routes/s3Router.ts
