+++ src/redux/reducers/methodsReducer.ts
