+++ src/redux/sagas/updateProfileSaga.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { put, takeEvery, call } from 'redux-saga/effects';
 import { SagaActions, SagaDispatch } from '../../models/sagaResource';
-import { UpdateProfilePayload } from '../../models/payloadResource';
+import { UpdateProfilePayload } from '../../models/sagaPayloadResource';
 
 // PUT route to update information for both a new and existing user
 // Since registration creates a username and password but nothing else,
